requirements/testproj.txt
# test project requirements
Pillow>=4.3.0
django-filter
djangorestframework-camel-case>=1.1.2
djangorestframework-recursive>=0.1.2
dj-database-url>=0.4.2
user_agents>=1.1.0
django-cors-headers
django-oauth-toolkit

